moving taskedit to views folder
